vpr: Move connection-based routing code into separate file
